LeetCodeLibrary/LeetCodeLibrary/TwoSum.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace LeetCodeLibrary
{
    public class TwoSum
    {
        public int[] FindSum(int[] nums, int target)
        {
            Dictionary<int, int> map = new Dictionary<int, int>();

            for (int i = 0; i < nums.Length; i++)
            {
                int complement = target - nums[i];
                Console.WriteLine($"Index: {i}, Num: {nums[i]}, Complement: {complement}");

                if (map.ContainsKey(complement))
                {
                    return new int[] { map[complement], i };
                }
                map[nums[i]] = i;
            }

            return new int[0];
        }
    }
}
